test: utils tests
