+++ lib/parsers.ts
@@ -2,88 +2,18 @@
  * Created by theophy on 02/08/2017.
  */
 import * as path from 'path';
-import { SwaggerSailsModel, SwaggerAttribute, SwaggerAction } from './interfaces';
-import swaggerJSDoc from 'swagger-jsdoc';
+import { SwaggerSailsModel, SwaggerRouteInfo } from './interfaces';
 import cloneDeep from 'lodash/cloneDeep';
-import mapKeys from 'lodash/mapKeys';
-import get from 'lodash/get';
-import uniqByLodash from 'lodash/uniqBy';
-import uniq from 'lodash/uniq';
 import { OpenApi } from '../types/openapi';
+import { generateSwaggerPath } from './generators';
+import { loadSwaggerDocComments, mergeSwaggerSpec, mergeSwaggerPaths } from './utils';
 
 declare const sails: any;
 
-const uniqBy = uniqByLodash as unknown as <T>(array: Array<T>, prop: string) => Array<T>
-
-const componentReference: Array<keyof OpenApi.Components> = [
-  "schemas",
-  "parameters",
-  "securitySchemes",
-  "requestBodies",
-  "responses",
-  "headers",
-  "examples",
-  "links",
-  "callbacks"
-]
 
 // consider all models except associative tables and 'Archive' model special case
 const removeModelExceptions = (model: SwaggerSailsModel): boolean => !!model.globalId && model.globalId !== 'Archive'
 
-export const mergeSwaggerSpec = (destination: SwaggerAttribute, source: SwaggerAttribute): SwaggerAttribute => {
-  const dest: SwaggerAttribute = cloneDeep(destination)
-  dest.tags = uniqBy([
-    ...(get(source, 'tags', [])),
-    ...(get(dest, 'tags', [])),
-  ], 'name');
-
-  dest.components = componentReference.reduce((mergedComponents, ref) => {
-    const type = get(source, `components.${ref}`);
-    if (type) {
-      mergedComponents[ref] = {
-        ...(get(dest.components, ref, {})),
-        ...type
-      }
-    }
-    return mergedComponents
-  }, dest.components || {})
-  return dest
-}
-
-export const mergeSwaggerPaths = (dest: SwaggerAction | undefined, sourcePaths: SwaggerAction): SwaggerAction => {
-  // strip off leading '/' from both source and destination swagger action, swagger-doc returns paths with leading /path-name: {}
-  const swaggerActions = mapKeys(cloneDeep(dest || {}), (_, key) => key.replace(/^\//, ''));
-  const paths = mapKeys(cloneDeep(sourcePaths), (_, key) => key.replace(/^\//, ''));
-  // unique action keys
-  const actions = uniq(Object.keys(swaggerActions).concat(Object.keys(paths)))
-  return actions.reduce((merged, key) => {
-    const destination = swaggerActions[key];
-    const source = paths[key];
-    merged[key] = {
-      ...(destination || {}),
-      ...(source || {})
-    }
-    return merged
-  }, {} as SwaggerAction)
-}
-
-export const loadSwaggerDocComments = (filePath: string): Promise<OpenApi.OpenApi>=> {
-  return new Promise((resolve, reject) => {
-    try {
-      const opts = {
-        definition: {
-          openapi: '3.0.0', // Specification (optional, defaults to swagger: '2.0')
-          info: { title: 'dummy', version: '0.0.0' },
-        },
-        apis: [filePath],
-      };
-      const specification = swaggerJSDoc(opts);
-      resolve(specification as OpenApi.OpenApi)
-    }catch(err){
-      reject(err)
-    }
-  });
-}
 
 export const parseModels = async (sailsConfig: Sails.Config, sailsModels: Array<SwaggerSailsModel> = []): Promise<{[globalId: string]: SwaggerSailsModel}> => {
   const models = sailsModels.filter(removeModelExceptions);
@@ -123,4 +53,42 @@ export const parseModels = async (sailsConfig: Sails.Config, sailsModels: Array<
       parsed[model.globalId] = model;
       return parsed
     }, {} as { [globalId: string]: SwaggerSailsModel })
+}
+
+export const parseCustomRoutes = (sailsConfig: Sails.Config): Array<SwaggerRouteInfo>  => {
+  for(const routeAddress in sailsConfig.routes){
+    // Parse 1: Route Address
+    // see https://sailsjs.com/documentation/concepts/routes/custom-routes#?route-address
+    let [verb, path] = routeAddress.split(/\s+/)
+    if (!path) {
+      path = verb
+      verb = 'all'
+    }
+
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?wildcards-and-dynamic-parameters
+    const swaggerPathParams = generateSwaggerPath(path);
+
+    // XXX TODO wildcard
+
+    // XXX TODO Handle arrays
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?route-target
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?policy-target-syntax
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?function-target-syntax
+
+    // XXX TODO Handle these variations
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?regular-expressions-in-addresses
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?routing-to-blueprint-actions
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?redirect-target-syntax (starts with / or scheme)
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?response-target-syntax
+    // https://sailsjs.com/documentation/concepts/routes/custom-routes#?policy-target-syntax
+    // https://sailsjs.com/documentation/concepts/actions-and-controllers#?actions-2
+
+    // Parse 2: Route target
+    const routerTarget = sailsConfig.routes[routeAddress];
+    // if(typeof routerTarget === 'string') {
+      
+    // }
+
+  }
+   return [] 
 }
